Box2D limitation: you can't run physics and collision at the same time. Collision is required for touch authority. This means the server has to have collision to distribute authority.
A big advantage of the physics is thus removed from the Distributed Authority System, the only advantage at this moment is that the player won't experience lag when moving an object.

I want to make the server chunkowner of every chunk, but I'm not sure if I will keep the Authority System instead of just using client prediction.